requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
pyyaml
